+++ Model/Login.cs
@@ -12,7 +12,7 @@ namespace BlackBox.Model
         public string FechaVersion { get; set; }
         public string Sitio { get; set; }
         public string Computadora { get; set; }
-        public double Taza { get; set; }
+        public decimal Taza { get; set; }
         public string Cajero { get; set; }
     }
 }
